refactor: ratatui::init and ratatui::restore
