+++ examples/minimal.rs
@@ -16,7 +16,6 @@
 use ratatui::{
     crossterm::event::{self, Event, KeyCode, KeyEventKind},
     text::Text,
-    Terminal,
 };
 
 /// This is a bare minimum example. There are many approaches to running an application loop, so
@@ -25,8 +24,8 @@ use ratatui::{
 ///
 /// [examples]: https://github.com/ratatui-org/ratatui/blob/main/examples
 /// [hello-world]: https://github.com/ratatui-org/ratatui/blob/main/examples/hello_world.rs
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let mut terminal = Terminal::init()?;
+fn main() -> std::io::Result<()> {
+    let mut terminal = ratatui::try_init()?;
     loop {
         terminal.draw(|frame| frame.render_widget(Text::raw("Hello World!"), frame.area()))?;
         if let Event::Key(key) = event::read()? {
@@ -35,6 +34,5 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             }
         }
     }
-    Terminal::restore()?;
-    Ok(())
+    ratatui::try_restore()
 }
