+++ src/terminal.rs
@@ -32,9 +32,13 @@
 //! [`Buffer`]: crate::buffer::Buffer
 
 mod frame;
+#[cfg(feature = "crossterm")]
+mod init;
 mod terminal;
 mod viewport;
 
 pub use frame::{CompletedFrame, Frame};
+#[cfg(feature = "crossterm")]
+pub use init::{init, restore, try_init, try_restore, DefaultTerminal};
 pub use terminal::{Options as TerminalOptions, Terminal};
 pub use viewport::Viewport;
